Move dist functions into dist_utils.py
